+++ R/select_variants.R
@@ -133,7 +133,7 @@ select_variants = function(cesa, genes = NULL, min_freq = 0, variant_ids = NULL,
                                                            end.field = "pos", seqinfo = GenomeInfoDb::seqinfo(final_gr))
     captured = cesa@mutations$sbs[IRanges::overlapsAny(query = mutations_gr, subject = final_gr, type = "within")]
     if (captured[, .N] == 0) {
-      stop("No mutations captured by input genomic positions (gr/variant_position_table).", call. = F)
+      return(data.table())
     }
     selected_sbs_ids = intersect(selected_sbs_ids, captured$sbs_id)
     aac_passing_gr = cesa@mutations$aac_sbs_key[captured$sbs_id, unique(aac_id), on = 'sbs_id', nomatch = NULL]